Move add project and edit project pages to /profile/sideProjectHub
Previously they were in /resources/sideProjectHub
