+++ components/studentPortal/profileV2/sideProjectHub/addProjectComponent.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
-import { Button } from "../../ui/Button";
-import { Input } from "../../ui/Input";
+import { Button } from "../../../ui/Button";
+import { Input } from "../../../ui/Input";
 
 export default function AddProjectComponent() {
   const [projectName, setProjectName] = useState("");
