+++ pages/profile/sideProjectHub/addProject/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import AddProjectComponent from "../../../../components/studentPortal/profileV2/sideProjectHub/addProjectComponent";
+
+export default function AddProjectPage() {
+  return <AddProjectComponent />;
+}
+
+AddProjectPage.auth = true;
